Implement resolve_multiple (closes #3, #5, #6, #7)
